langgraph_rag_demo/requirements.txt
langgraph
langchain
langchain-community
faiss-cpu
sentence-transformers
transformers
accelerate
beautifulsoup4
requests
python-dotenv
